fix(react-typescript-components): rename enum to constants
