+++ packages/react-typescript-components/src/button/components/text-button/index.tsx
@@ -2,9 +2,8 @@ import { type FC, type ReactElement, useMemo } from 'react'
 import clsx from 'clsx'
 // components
 import { P1, P2 } from '../../../text/paragraph'
-// enums
-import { SIZE, type Size, STYLE, type Style } from '../../constants'
 // constants
+import { SIZE, type Size, STYLE, type Style } from '../../constants'
 import { THEME, type Theme } from '../../../constants/theme'
 // utils
 import {
